Revert "HTML emails support"
