+++ IntraWeb/src/IntraWeb/appsettings.json
@@ -4,22 +4,6 @@
       "ConnectionString": "Server=(localdb)\\mssqllocaldb;Database=aspnet5-IntraWeb-674f6dea-b458-402b-8968-06bd3dab3430;Trusted_Connection=True;MultipleActiveResultSets=true"
     }
   },
-  "Email": {
-    "Gmail_SMTP": {
-      "serverAddress": "smtp.gmail.com",
-      "port": "587",
-      "useSsl": true,
-      "userName": "intrawebkros@gmail.com",
-      "password": "21IntraWeb12"
-    },
-    "SendGrid_SMTP": {
-      "serverAddress": "smtp.sendgrid.net",
-      "port": "587",
-      "useSsl": false,
-      "userName": "azure_c79ddf93ae96f676b0cf33bb83ab7d8d@azure.com",
-      "password": "SrqNLgW60xDiNMI"
-    }
-  },
   "Logging": {
     "IncludeScopes": false,
     "LogLevel": {
