+++ IntraWeb/src/IntraWeb/project.json
@@ -8,7 +8,6 @@
     "AutoMapper": "4.1.1",
     "EntityFramework.Commands": "7.0.0-rc1-final",
     "EntityFramework.MicrosoftSqlServer": "7.0.0-rc1-final",
-    "MailKit": "1.3.0-beta5",
     "Microsoft.AspNet.Authentication.Cookies": "1.0.0-rc1-final",
     "Microsoft.AspNet.Diagnostics.Entity": "7.0.0-rc1-final",
     "Microsoft.AspNet.Identity.EntityFramework": "3.0.0-rc1-final",
@@ -32,10 +31,7 @@
     "ef": "EntityFramework.Commands"
   },
   "frameworks": {
-    "dnx451": {
-      "dependencies": {
-      }
-    },
+    "dnx451": { },
     "dnxcore50": { }
   },
   "exclude": [
@@ -46,6 +42,9 @@
     "**.user",
     "**.vspscc"
   ],
+  "scripts": {
+    "prepublish": [ "npm install", "bower install", "gulp clean", "gulp min" ]
+  },
   "configurations": {
     "Testing": { }
   }
