+++ IntraWeb/src/IntraWeb/wwwroot/app/boot.js
