+++ IntraWeb/src/IntraWeb/wwwroot/lib/angular2-polyfills.js
